+++ src/Clinica Frba/Clases/Funcionalidad.cs
@@ -7,7 +7,7 @@ namespace Clinica_Frba.Clases
 {
     class Funcionalidad
     {
-        public int Id { get; set; }
+        public decimal Id { get; set; }
         public string Nombre { get; set; }
     }
 }
